requirements.txt
google-genai>=1.0.0
openai>=1.0.0
streamlit>=1.45.0
Pillow>=10.0.0
plotly>=5.24.0
pydantic>=2.10.0
python-dotenv>=1.0.0
jinja2>=3.1.0
reportlab>=4.0.0
pymupdf>=1.24.0
